+++ core/server/api/v2/redirects.js
@@ -1,5 +1,5 @@
 const web = require('../../web');
-const redirects = require('../../../frontend/services/redirects');
+const redirects = require('../../services/redirects');
 
 module.exports = {
     docName: 'redirects',
