+++ core/server/web/shared/middlewares/custom-redirects.js
@@ -7,7 +7,7 @@ const urlUtils = require('../../../../shared/url-utils');
 const errors = require('@tryghost/errors');
 const i18n = require('../../../../shared/i18n');
 const logging = require('@tryghost/logging');
-const redirectsService = require('../../../../frontend/services/redirects');
+const redirectsService = require('../../../services/redirects');
 
 const _private = {};
